requirements.txt
bcrypt==4.3.0
boto3==1.38.0
botocore==1.38.0
certifi
dnspython==2.7.0
jmespath==1.0.1
passlib==1.7.4
pyjwt==2.10.1
pymongo==4.12.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
s3transfer==0.12.0
six==1.17.0
urllib3>=1.26.0,<2.0.0
pydantic-settings

# Testing
pytest
pytest-mock
requests
